refactor: group all data libs and remove old ones
