+++ packages/core/components/DropZone/lib/use-content-with-preview.ts
@@ -1,7 +1,7 @@
 import { Preview } from "./../context";
 import { useEffect, useState } from "react";
 import { useRenderedCallback } from "../../../lib/dnd/use-rendered-callback";
-import { insert } from "../../../lib/insert";
+import { insert } from "../../../lib/data/insert";
 import { ZoneStoreContext } from "../context";
 import { useContextStore } from "../../../lib/use-context-store";
 import { useAppStore } from "../../../store";
