+++ packages/core/lib/data/find-zones-for-area.ts
@@ -1,4 +1,4 @@
-import { PrivateAppState } from "../types/Internal";
+import { PrivateAppState } from "../../types/Internal";
 
 export const findZonesForArea = (state: PrivateAppState, area: string) => {
   return Object.keys(state.indexes.zones).filter(
