+++ packages/core/lib/data/for-related-zones.ts
@@ -1,5 +1,5 @@
-import { Content, Data } from "../types";
-import { getZoneId } from "./get-zone-id";
+import { Content, Data } from "../../types";
+import { getZoneId } from "../get-zone-id";
 
 export function forRelatedZones<UserData extends Data>(
   item: UserData["content"][0],
