+++ packages/core/lib/data/remove.ts
