+++ packages/core/lib/data/replace.ts
