+++ packages/core/lib/data/to-component.ts
@@ -1,4 +1,4 @@
-import { ComponentData, RootData } from "../types";
+import { ComponentData, RootData } from "../../types";
 
 export const toComponent = (item: ComponentData | RootData): ComponentData => {
   return "type" in item
