+++ packages/core/lib/index.ts
@@ -1,4 +1,4 @@
 export * from "./filter";
 export { default as getClassNameFactory } from "./get-class-name-factory";
-export * from "./reorder";
-export * from "./replace";
+export * from "./data/reorder";
+export * from "./data/replace";
