--- packages/core/lib/selector-is.ts
@@ -1,4 +1,0 @@
-import { ItemSelector } from "./get-item";
-
-export const selectorIs = (a: ItemSelector | null, b: ItemSelector | null) =>
-  a?.zone === b?.zone && a?.index === b?.index;
