+++ packages/core/lib/use-breadcrumbs.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 import { useAppStore, useAppStoreApi } from "../store";
-import { ItemSelector } from "./get-item";
+import { ItemSelector } from "./data/get-item";
 
 export type Breadcrumb = {
   label: string;
