+++ packages/core/reducer/actions/set.ts
@@ -2,7 +2,7 @@ import { Data } from "../../types";
 import { SetAction } from "../actions";
 import { AppStore } from "../../store";
 import { PrivateAppState } from "../../types/Internal";
-import { walkTree } from "../../lib/walk-tree";
+import { walkTree } from "../../lib/data/walk-tree";
 
 export const setAction = <UserData extends Data>(
   state: PrivateAppState<UserData>,
